add Prisma and get events from the db
