+++ .env.example
@@ -1,2 +1,3 @@
 NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY=
 CLERK_SECRET_KEY=
+DATABASE_URL=''
