+++ package.json
@@ -14,7 +14,9 @@
     "@fullcalendar/daygrid": "^6.1.10",
     "@fullcalendar/interaction": "^6.1.10",
     "@fullcalendar/react": "^6.1.10",
+    "@prisma/client": "^5.9.1",
     "next": "14.0.4",
+    "prisma": "^5.9.1",
     "react": "^18",
     "react-dom": "^18"
   },
